Check if scheme can be opened before opening
This commit adds a check to the helper that will retrieve
the path to the app that responds to the given scheme. Failure
to retrieve this path means the app does not exist on the
system and the scheme shouldn't be called.

It's a fix for a rare case when the container app has been
moved to the Trash, but the helper is still in launchd. The OS
will throw errors stating the scheme can't be opened.